Allow dynamic pin types to be correctly connected within the graph by checking if the referenced start and end pin has changed after being connected.

+++ Assets/Framework/NodeEditor/Core/CoreNodes.cs
@@ -52,8 +52,8 @@ namespace Framework.NodeSystem
             }
             else if (pinConnectEvent.OtherPin.WrappedType == typeof(int))
             {
-                In = ChangePinType<float>(In);
-                Out = ChangePinType<float>(Out);
+                In = ChangePinType<int>(In);
+                Out = ChangePinType<int>(Out);
             }
         }
